@@ -30,13 +30,13 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "element-ui": "^2.4.5",
-    "vue": "^2.6.14"
+    "vue": "^3.2.45",
+    "element-plus": "2.3.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.6.0",
+    "@vue/cli-plugin-babel": "^4.5.0",
     "@vue/cli-plugin-eslint": "^3.6.0",
-    "@vue/cli-service": "^3.6.0",
+    "@vue/cli-service": "^4.5.0",
     "@vue/eslint-config-standard": "^4.0.0",
     "babel-eslint": "^10.0.1",
     "babel-plugin-component": "^1.1.1",
@@ -45,7 +45,7 @@
     "sass": "^1.18.0",
     "sass-loader": "^7.1.0",
     "vue-cli-plugin-element": "^1.0.1",
-    "vue-template-compiler": "^2.5.21"
+    "@vue/compiler-sfc": "^3.1.1"
   },
   "eslintConfig": {
     "root": true,
